Changed ItemTagDTO to EditItemDTO and changed funtionality accordingly. Changed some styling. Not quite complete, but working.
