+++ src/main/java/org/launchcode/andrewgroupa/security/WebSecurity.java
@@ -20,35 +20,35 @@ import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 @EnableWebSecurity
 public class WebSecurity extends WebSecurityConfigurerAdapter {
 
-    @Autowired
-    UserDetailsService userDetailsService;
+  @Autowired
+  UserDetailsService userDetailsService;
 
 
-    @Override
-    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-        auth.userDetailsService(userDetailsService);
-    }
+  @Override
+  protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+    auth.userDetailsService(userDetailsService);
+  }
 
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http.authorizeRequests()
-                    .antMatchers("/list", "/tags", "/logout").permitAll()//.hasAnyRole("ADMIN", "USER")
-                    .antMatchers("/").permitAll()
-                    .and().formLogin()
-                    .loginPage("/login").permitAll();
+  @Override
+  protected void configure(HttpSecurity http) throws Exception {
+    http.authorizeRequests()
+        .antMatchers("/list", "/tags", "/edit", "/logout").permitAll()//.hasAnyRole("ADMIN", "USER")
+        .antMatchers("/").permitAll()
+        .and().formLogin()
+        .loginPage("/login").permitAll();
 
-        http.logout()
-                .logoutRequestMatcher(new AntPathRequestMatcher("/logout"))
-                .logoutSuccessUrl("/logout/done")
-                .invalidateHttpSession(true);
+    http.logout()
+        .logoutRequestMatcher(new AntPathRequestMatcher("/logout"))
+        .logoutSuccessUrl("/logout/done")
+        .invalidateHttpSession(true);
 
-        http.csrf().disable();
+    http.csrf().disable();
 
-    }
+  }
 
-    @Bean
-    public PasswordEncoder getPasswordEncoder(){
-        return new BCryptPasswordEncoder();
-    }
+  @Bean
+  public PasswordEncoder getPasswordEncoder() {
+    return new BCryptPasswordEncoder();
+  }
 
 }
