+++ src/main/resources/templates/fragments.html
@@ -8,12 +8,12 @@
         name="viewport">
   <title>Shopping List</title>
   <link crossorigin="anonymous"
-        href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
-        integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
+        href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.1/dist/css/bootstrap.min.css"
+        integrity="sha384-F3w7mX95PdgyTmZZMECAngseQB83DfGTowi0iMjiWaeVhAn4FJkqJByhZMI3AhiU"
         rel="stylesheet">
   <script crossorigin="anonymous"
-          integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
-          src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"></script>
+          integrity="sha384-/bQdsTh/da6pkI1MST/rWKFNjaCP5gBSY4sEBT38Q/9RBh9AH40zEOg7Hlq2THRZ"
+          src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.1/dist/js/bootstrap.bundle.min.js"></script>
   <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.5.0/font/bootstrap-icons.css"
         rel="stylesheet">
   <link rel="stylesheet"
@@ -44,21 +44,24 @@
       <ul class="navbar-nav mr-auto">
         <li class="nav-item">
           <a class="nav-link"
-             href="/list">My List</a>
+             href="/list">All Items</a>
         </li>
         <li class="nav-item">
           <a class="nav-link"
-             href="/list/shopping">My Lists</a>
+             href="/list/shopping"
+             sec:authorize="isAuthenticated()">My Lists</a>
         </li>
         <li class="nav-item">
           <a class="nav-link"
              href="/tags">Tags</a>
         </li>
-        <li class="nav-item">
+        <li class="nav-item"
+            sec:authorize="isAnonymous()">
           <a class="nav-link"
              href="/login">Login</a>
         </li>
-        <li class="nav-item">
+        <li class="nav-item"
+            sec:authorize="isAuthenticated()">
           <a class="nav-link"
              href="/logout">Logout</a>
         </li>
