+++ src/main/resources/templates/list/detail.html
@@ -8,9 +8,7 @@
 </div>
 <div class="container">
   <div class="content">
-    <div>
-      <h2 th:text="'Current List: ' + ${currentList.name}"></h2>
-    </div>
+    <h2 th:text="'Current List: ' + ${currentList.name}"></h2>
     <div>
       <form class="d-flex"
             method="post">
@@ -21,7 +19,7 @@
                  th:field="${item.name}">
           <p class="error"
              th:errors="${item.name}"></p>
-          <button class="btn btn-dark d-flex w-25"
+          <button class="btn btn-dark w-25"
                   id="button-addon2"
                   type="submit">Add
           </button>
